RavenDB-8760 Move (etl, connection_strings, revision, expiration) endpoints to the right handlers
