+++ src/Raven.Client/ServerWide/Operations/ETL/ResetEtlOperation.cs
@@ -41,7 +41,7 @@ namespace Raven.Client.ServerWide.Operations.ETL
 
             public override HttpRequestMessage CreateRequest(JsonOperationContext ctx, ServerNode node, out string url)
             {
-                url = $"{node.Url}/admin/etl?configuration-name={_configurationName}&transformation-name={_transformationName}&name={_databaseName}";
+                url = $"{node.Url}/databases/{_databaseName}/admin/etl?configuration-name={_configurationName}&transformation-name={_transformationName}";
 
                 var request = new HttpRequestMessage
                 {
