+++ src/Raven.Server/Web/RequestHandler.cs
@@ -494,58 +494,5 @@ namespace Raven.Server.Web
                 writer.Flush();
             }
         }
-
-        protected async Task DatabaseConfigurations(Func<TransactionOperationContext, string,
-           BlittableJsonReaderObject, Task<(long, object)>> setupConfigurationFunc,
-           string debug,
-           Func<string, BlittableJsonReaderObject, bool> beforeSetupConfiguration = null,
-           Action<DynamicJsonValue, BlittableJsonReaderObject, long> fillJson = null, string databaseName = null, 
-           HttpStatusCode statusCode = HttpStatusCode.OK)
-        {
-            var name = databaseName ?? GetQueryStringValueAndAssertIfSingleAndNotEmpty("name");
-
-            if (TryGetAllowedDbs(name, out var _, requireAdmin: true) == false)
-                return;
-
-            if (ResourceNameValidator.IsValidResourceName(name, ServerStore.Configuration.Core.DataDirectory.FullPath, out string errorMessage) == false)
-                throw new BadRequestException(errorMessage);
-
-            ServerStore.EnsureNotPassive();
-            using (ServerStore.ContextPool.AllocateOperationContext(out TransactionOperationContext context))
-            {
-                var configurationJson = await context.ReadForMemoryAsync(RequestBodyStream(), debug);
-                if (beforeSetupConfiguration?.Invoke(name, configurationJson) == false)
-                    return;
-
-                var (index, _) = await setupConfigurationFunc(context, name, configurationJson);
-                DatabaseRecord dbRecord;
-                using (context.OpenReadTransaction())
-                {
-                    //TODO: maybe have a timeout here for long loading operations
-                    dbRecord = ServerStore.Cluster.ReadDatabase(context, name);
-                }
-                if (dbRecord.Topology.RelevantFor(ServerStore.NodeTag))
-                {
-                    var db = await ServerStore.DatabasesLandlord.TryGetOrCreateResourceStore(name);
-                    await db.RachisLogIndexNotifications.WaitForIndexNotification(index);
-                }
-                else
-                {
-                    await ServerStore.Cluster.WaitForIndexNotification(index);
-                }
-                HttpContext.Response.StatusCode = (int)statusCode;
-
-                using (var writer = new BlittableJsonTextWriter(context, ResponseBodyStream()))
-                {
-                    var json = new DynamicJsonValue
-                    {
-                        ["RaftCommandIndex"] = index
-                    };
-                    fillJson?.Invoke(json, configurationJson, index);
-                    context.Write(writer, json);
-                    writer.Flush();
-                }
-            }
-        }
     }
 }
